Backend/Document_Service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.2